Hinstop/requirements.txt
indic-nlp-library